hv_test_system_project_v20/hv_test_system/common.py
from __future__ import annotations

import sys
import serial
import serial.tools.list_ports
import struct
import threading
import time
import math
from datetime import datetime
from collections import deque
from queue import Queue, Empty
from PyQt5 import QtGui
from PyQt5.QtWidgets import *
from PyQt5.QtCore import QTimer, QThread, pyqtSignal, Qt, QMutex, QMutexLocker, QSize, QEventLoop
from PyQt5.QtGui import QFont, QColor
import pyqtgraph as pg
import numpy as np
import warnings
import gc
import traceback
import configparser
import os
import csv
import pathlib

# 抑制各种警告
from .constants import CONFIG_FILE, TEMP_DATA_FILE, DATA_HEADERS, DATA_HEADER_LINE
from .logging_utils import setup_logger

logger = setup_logger()

warnings.filterwarnings("ignore")

# 设置pyqtgraph使用抗锯齿
pg.setConfigOptions(antialias=True)





