hv_test_system_project_v20/requirements.txt
pyserial
PyQt5
pyqtgraph
numpy
pyvisa
fastapi
uvicorn[standard]
requests
